Preferring to version the Graphql controller. Enabled openapi validator for graphql
